+++ pkg/controller/mongodb/replicaset_controller_test.go
@@ -7,8 +7,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/mongodb/mongodb-kubernetes-operator/pkg/kube/configmap"
-
 	k8sClient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/mongodb/mongodb-kubernetes-operator/pkg/authentication/scram"
@@ -355,189 +353,6 @@ func TestScramIsConfigured(t *testing.T) {
 		assert.NoError(t, err)
 		assert.Equal(t, s.Data[scram.AgentKeyfileKey], []byte(currentAc.Auth.Key))
 		assert.Equal(t, s.Data[scram.AgentPasswordKey], []byte(currentAc.Auth.AutoPwd))
-	})
-}
-
-func TestStatefulSet_IsCorrectlyConfiguredWithTLS(t *testing.T) {
-	mdb := newTestReplicaSetWithTLS()
-	mgr := client.NewManager(&mdb)
-
-	s := secret.Builder().
-		SetName(mdb.Spec.Security.TLS.CertificateKeySecret.Name).
-		SetNamespace(mdb.Namespace).
-		SetField("tls.crt", "CERT").
-		SetField("tls.key", "KEY").
-		Build()
-
-	err := mgr.GetClient().Create(context.TODO(), &s)
-	assert.NoError(t, err)
-
-	configMap := configmap.Builder().
-		SetName(mdb.Spec.Security.TLS.CaConfigMap.Name).
-		SetNamespace(mdb.Namespace).
-		SetField("ca.crt", "CERT").
-		Build()
-
-	err = mgr.GetClient().Create(context.TODO(), &configMap)
-	assert.NoError(t, err)
-
-	r := newReconciler(mgr, mockManifestProvider(mdb.Spec.Version))
-	res, err := r.Reconcile(reconcile.Request{NamespacedName: types.NamespacedName{Namespace: mdb.Namespace, Name: mdb.Name}})
-	assertReconciliationSuccessful(t, res, err)
-
-	sts := appsv1.StatefulSet{}
-	err = mgr.GetClient().Get(context.TODO(), types.NamespacedName{Name: mdb.Name, Namespace: mdb.Namespace}, &sts)
-	assert.NoError(t, err)
-
-	// Assert that all TLS volumes have been added.
-	assert.Len(t, sts.Spec.Template.Spec.Volumes, 6)
-	assert.Contains(t, sts.Spec.Template.Spec.Volumes, corev1.Volume{
-		Name: "tls",
-		VolumeSource: corev1.VolumeSource{
-			EmptyDir: &corev1.EmptyDirVolumeSource{},
-		},
-	})
-	assert.Contains(t, sts.Spec.Template.Spec.Volumes, corev1.Volume{
-		Name: "tls-ca",
-		VolumeSource: corev1.VolumeSource{
-			ConfigMap: &corev1.ConfigMapVolumeSource{
-				LocalObjectReference: corev1.LocalObjectReference{
-					Name: mdb.Spec.Security.TLS.CaConfigMap.Name,
-				},
-			},
-		},
-	})
-	assert.Contains(t, sts.Spec.Template.Spec.Volumes, corev1.Volume{
-		Name: "tls-secret",
-		VolumeSource: corev1.VolumeSource{
-			Secret: &corev1.SecretVolumeSource{
-				SecretName: mdb.Spec.Security.TLS.CertificateKeySecret.Name,
-			},
-		},
-	})
-
-	// Assert that the TLS init container has been added.
-	tlsVolumeMount := corev1.VolumeMount{
-		Name:      "tls",
-		ReadOnly:  false,
-		MountPath: tlsServerMountPath,
-	}
-	tlsSecretVolumeMount := corev1.VolumeMount{
-		Name:      "tls-secret",
-		ReadOnly:  true,
-		MountPath: tlsSecretMountPath,
-	}
-	tlsCAVolumeMount := corev1.VolumeMount{
-		Name:      "tls-ca",
-		ReadOnly:  true,
-		MountPath: tlsCAMountPath,
-	}
-
-	assert.Len(t, sts.Spec.Template.Spec.InitContainers, 2)
-	tlsInitContainer := sts.Spec.Template.Spec.InitContainers[1]
-	assert.Equal(t, "tls-init", tlsInitContainer.Name)
-
-	// Assert that all containers have the correct volumes mounted.
-	assert.Len(t, tlsInitContainer.VolumeMounts, 2)
-	assert.Contains(t, tlsInitContainer.VolumeMounts, tlsVolumeMount)
-	assert.Contains(t, tlsInitContainer.VolumeMounts, tlsSecretVolumeMount)
-
-	agentContainer := sts.Spec.Template.Spec.Containers[0]
-	assert.Contains(t, agentContainer.VolumeMounts, tlsVolumeMount)
-	assert.Contains(t, agentContainer.VolumeMounts, tlsCAVolumeMount)
-
-	mongodbContainer := sts.Spec.Template.Spec.Containers[1]
-	assert.Contains(t, mongodbContainer.VolumeMounts, tlsVolumeMount)
-	assert.Contains(t, mongodbContainer.VolumeMounts, tlsCAVolumeMount)
-}
-
-func TestAutomationConfig_IsCorrectlyConfiguredWithTLS(t *testing.T) {
-	createAC := func(mdb mdbv1.MongoDB) automationconfig.AutomationConfig {
-		manifest, err := mockManifestProvider(mdb.Spec.Version)()
-		assert.NoError(t, err)
-		versionConfig := manifest.BuildsForVersion(mdb.Spec.Version)
-
-		ac, err := buildAutomationConfig(
-			mdb,
-			versionConfig,
-			automationconfig.AutomationConfig{},
-			getTLSConfigModification(mdb),
-		)
-		assert.NoError(t, err)
-		return ac
-	}
-
-	t.Run("With TLS disabled", func(t *testing.T) {
-		mdb := newTestReplicaSet()
-		ac := createAC(mdb)
-
-		assert.Equal(t, automationconfig.TLS{
-			CAFilePath:            "",
-			ClientCertificateMode: automationconfig.ClientCertificateModeOptional,
-		}, ac.TLS)
-
-		for _, process := range ac.Processes {
-			assert.Equal(t, automationconfig.MongoDBTLS{
-				Mode: automationconfig.TLSModeDisabled,
-			}, process.Args26.Net.TLS)
-		}
-	})
-
-	t.Run("With TLS enabled, during rollout", func(t *testing.T) {
-		mdb := newTestReplicaSetWithTLS()
-		ac := createAC(mdb)
-
-		assert.Equal(t, automationconfig.TLS{
-			CAFilePath:            "",
-			ClientCertificateMode: automationconfig.ClientCertificateModeOptional,
-		}, ac.TLS)
-
-		for _, process := range ac.Processes {
-			assert.Equal(t, automationconfig.MongoDBTLS{
-				Mode: automationconfig.TLSModeDisabled,
-			}, process.Args26.Net.TLS)
-		}
-	})
-
-	t.Run("With TLS enabled and required, rollout completed", func(t *testing.T) {
-		mdb := newTestReplicaSetWithTLS()
-		mdb.Annotations[tLSRolledOutAnnotationKey] = "true"
-		ac := createAC(mdb)
-
-		assert.Equal(t, automationconfig.TLS{
-			CAFilePath:            tlsCAMountPath + tlsCACertName,
-			ClientCertificateMode: automationconfig.ClientCertificateModeOptional,
-		}, ac.TLS)
-
-		for _, process := range ac.Processes {
-			assert.Equal(t, automationconfig.MongoDBTLS{
-				Mode:                               automationconfig.TLSModeRequired,
-				PEMKeyFile:                         tlsServerMountPath + tlsServerFileName,
-				CAFile:                             tlsCAMountPath + tlsCACertName,
-				AllowConnectionsWithoutCertificate: true,
-			}, process.Args26.Net.TLS)
-		}
-	})
-
-	t.Run("With TLS enabled and optional, rollout completed", func(t *testing.T) {
-		mdb := newTestReplicaSetWithTLS()
-		mdb.Annotations[tLSRolledOutAnnotationKey] = "true"
-		mdb.Spec.Security.TLS.Optional = true
-		ac := createAC(mdb)
-
-		assert.Equal(t, automationconfig.TLS{
-			CAFilePath:            tlsCAMountPath + tlsCACertName,
-			ClientCertificateMode: automationconfig.ClientCertificateModeOptional,
-		}, ac.TLS)
-
-		for _, process := range ac.Processes {
-			assert.Equal(t, automationconfig.MongoDBTLS{
-				Mode:                               automationconfig.TLSModePreferred,
-				PEMKeyFile:                         tlsServerMountPath + tlsServerFileName,
-				CAFile:                             tlsCAMountPath + tlsCACertName,
-				AllowConnectionsWithoutCertificate: true,
-			}, process.Args26.Net.TLS)
-		}
 	})
 }
